@@ -1,16 +1,38 @@
 module github.com/popura-network/Popura
 
-go 1.16
+go 1.17
 
 require (
 	github.com/cheggaaa/pb/v3 v3.0.8
 	github.com/gologme/log v1.2.0
 	github.com/hashicorp/go-syslog v1.0.0
 	github.com/hjson/hjson-go v3.1.0+incompatible
-	github.com/kardianos/minwinsvc v1.0.0
-	github.com/miekg/dns v1.1.41 // indirect
+	github.com/kardianos/minwinsvc v1.0.2
 	github.com/mitchellh/mapstructure v1.4.1
-	github.com/yggdrasil-network/yggdrasil-go v0.4.3
+	github.com/olekukonko/tablewriter v0.0.5
+	github.com/yggdrasil-network/yggdrasil-go v0.4.6
 	github.com/zhoreeq/meshname v0.2.0
-	golang.org/x/text v0.3.8-0.20211004125949-5bd84dd9b33b
+	golang.org/x/mobile v0.0.0-20221012134814-c746ac228303
+	golang.org/x/text v0.3.8
+)
+
+require (
+	github.com/Arceliar/ironwood v0.0.0-20221025225125-45b4281814c2 // indirect
+	github.com/Arceliar/phony v0.0.0-20210209235338-dde1a8dca979 // indirect
+	github.com/VividCortex/ewma v1.2.0 // indirect
+	github.com/fatih/color v1.12.0 // indirect
+	github.com/mattn/go-colorable v0.1.8 // indirect
+	github.com/mattn/go-isatty v0.0.13 // indirect
+	github.com/mattn/go-runewidth v0.0.13 // indirect
+	github.com/miekg/dns v1.1.41 // indirect
+	github.com/rivo/uniseg v0.2.0 // indirect
+	github.com/vishvananda/netlink v1.1.0 // indirect
+	github.com/vishvananda/netns v0.0.0-20210104183010-2eb08e3e575f // indirect
+	golang.org/x/crypto v0.0.0-20221012134737-56aed061732a // indirect
+	golang.org/x/mod v0.6.0-dev.0.20220419223038-86c51ed26bb4 // indirect
+	golang.org/x/net v0.0.0-20221014081412-f15817d10f9b // indirect
+	golang.org/x/sys v0.0.0-20221013171732-95e765b1cc43 // indirect
+	golang.org/x/tools v0.1.12 // indirect
+	golang.zx2c4.com/wireguard v0.0.0-20211017052713-f87e87af0d9a // indirect
+	golang.zx2c4.com/wireguard/windows v0.4.12 // indirect
 )
